4.21 vain oikea käyttäjä saa poistettua blogin

+++ blogilista/controllers/blogs.js
@@ -35,6 +35,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if ( ! (blog.user.toString() === decodedToken.id.toString()) ) {
+    return response.status(401).json({ error: 'user not authorized to delete' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
